sparklin/JsonReceiverFuncApp/requirements.txt
azure-functions
azure-storage-blob
azure-data-tables
